+++ pkg/tracker/options_test.go
@@ -24,21 +24,21 @@ func TestOptions_validate(t *testing.T) {
 	options = Options{
 		URL:  "https://example.com/foo/bar?query=parameter#anchor",
 		Path: "/new/path",
+		Tags: map[string]string{
+			"key0":   "value0",
+			" key1 ": " value1 ",
+			"ignore": "",
+			"":       "ignore",
+		},
 	}
 	options.validate(req)
 	assert.Equal(t, "https://example.com/new/path?query=parameter#anchor", options.URL)
 	assert.Equal(t, "example.com", options.Hostname)
-}
-
-func TestOptionsFromRequest(t *testing.T) {
-	req := httptest.NewRequest(http.MethodGet, "https://foo.bar?url=https://example.com/test&t=Title&ref=Referrer&w=1920&h=1080", nil)
-	options := OptionsFromRequest(req)
-	options.validate(req)
-	assert.Equal(t, "https://example.com/test", options.URL)
-	assert.Equal(t, "example.com", options.Hostname)
-	assert.Equal(t, "/test", options.Path)
-	assert.Equal(t, "Title", options.Title)
-	assert.Equal(t, "Referrer", options.Referrer)
-	assert.Equal(t, uint16(1920), options.ScreenWidth)
-	assert.Equal(t, uint16(1080), options.ScreenHeight)
+	k, v := options.getTags()
+	assert.Len(t, k, 2)
+	assert.Len(t, v, 2)
+	assert.Equal(t, "key0", k[0])
+	assert.Equal(t, "key1", k[1])
+	assert.Equal(t, "value0", v[0])
+	assert.Equal(t, "value1", v[1])
 }